Extracted implicit defs into their own syntax file to reduce dependency to the WalletApi object
