+++ brambl-sdk/src/main/scala/co/topl/brambl/wallet/WalletApi.scala
@@ -18,6 +18,7 @@ import com.google.protobuf.ByteString
 import quivr.models._
 import quivr.models.VerificationKey._
 import quivr.models.SigningKey._
+import co.topl.brambl.syntax.{cryptoToPbKeyPair, cryptoVkToPbVk, pbKeyPairToCryptoKeyPair, pbVkToCryptoVk}
 import cats.implicits._
 
 import scala.language.implicitConversions
@@ -440,45 +441,6 @@ object WalletApi {
         )
       } yield res
     }
-  }
-
-  implicit def pbVkToCryptoVk(pbVk: VerificationKey.ExtendedEd25519Vk): ExtendedEd25519.PublicKey =
-    ExtendedEd25519.PublicKey(
-      signing.Ed25519.PublicKey(pbVk.vk.value.toByteArray),
-      pbVk.chainCode.toByteArray
-    )
-
-  implicit def pbKeyPairToCryotoKeyPair(
-    pbKeyPair: KeyPair
-  ): signing.KeyPair[ExtendedEd25519.SecretKey, ExtendedEd25519.PublicKey] =
-    signing.KeyPair(
-      ExtendedEd25519.SecretKey(
-        pbKeyPair.sk.sk.extendedEd25519.get.leftKey.toByteArray,
-        pbKeyPair.sk.sk.extendedEd25519.get.rightKey.toByteArray,
-        pbKeyPair.sk.sk.extendedEd25519.get.chainCode.toByteArray
-      ),
-      pbKeyPair.vk.vk.extendedEd25519.get
-    )
-
-  implicit def cryptoVkToPbVk(cryptoVk: ExtendedEd25519.PublicKey): VerificationKey.ExtendedEd25519Vk =
-    ExtendedEd25519Vk(
-      Ed25519Vk(ByteString.copyFrom(cryptoVk.vk.bytes)),
-      ByteString.copyFrom(cryptoVk.chainCode)
-    )
-
-  implicit def cryptoToPbKeyPair(
-    keyPair: signing.KeyPair[ExtendedEd25519.SecretKey, ExtendedEd25519.PublicKey]
-  ): KeyPair = {
-    val skCrypto = keyPair.signingKey
-    val sk = ExtendedEd25519Sk(
-      ByteString.copyFrom(skCrypto.leftKey),
-      ByteString.copyFrom(skCrypto.rightKey),
-      ByteString.copyFrom(skCrypto.chainCode)
-    )
-    KeyPair(
-      VerificationKey(VerificationKey.Vk.ExtendedEd25519(keyPair.verificationKey)),
-      SigningKey(SigningKey.Sk.ExtendedEd25519(sk))
-    )
   }
 
   case class NewWalletResult[F[_]](mnemonic: IndexedSeq[String], mainKeyVaultStore: VaultStore[F])
